==> 1_getHistoricalData.ipynb <==
import boto3
import pandas as pd
from datetime import datetime
import re

s3 = boto3.client('s3')

def get_all_s3_keys(bucket):
    """Get a list of all keys in an S3 bucket."""
    keysPlayers = []
    keysTournaments = []

    kwargs = {'Bucket': bucket}
    while True:
        resp = s3.list_objects_v2(**kwargs)
        for obj in resp['Contents']:
            if obj['Key'].split('/')[0] == 'tennis' and obj['Key'].split('/')[1] == "Players" and int(obj['Key'].split('/')[2]) == 2020:
                keysPlayers.append(obj['Key'])
            if obj['Key'].split('/')[0] == 'tennis' and obj['Key'].split('/')[1] == "Tournaments" and int(obj['Key'].split('/')[2]) == 2020:
                keysTournaments.append(obj['Key'])
                
        try:
            kwargs['ContinuationToken'] = resp['NextContinuationToken']
        except KeyError:
            break

    return keysPlayers, keysTournaments

keysPlayers, keysTournaments = get_all_s3_keys("pronoo")

all_tournaments = pd.DataFrame([], columns=['name', 'phase', 'surfaceInout', 'typeTournament', 'date', 'winnerRate',
       'looserRate', 'looserName', 'winnerName', 'winnerRateSurface',
       'looserRateSurface', 'url'])

all_players = pd.DataFrame([], columns=['player', 'aces', 'doubleFautes', 'firstServe', 'firstServePointsWon',
       'secondServePointsWon', 'breakPointsFaced', 'breakPointsSaved',
       'serviceGamePlayed', 'serviceGamesWon', 'totalServicePointsWon',
       'firstServeReturnPointsWon', 'secondServeReturnPointsWon',
       'breakPointsOpportunities', 'breakPointsConverted', 'returnGamesPlayed',
       'returnGamesWon', 'returnPointsWon', 'totalPointsWon', 'url', 'year',
       'weight', 'height', 'age', 'startCareer', 'plays', 'ranking', 'points',
       'tournCell', 'pointsDropping', 'nextBest'])

for i,key in enumerate(keysPlayers):
    if i%5000==0:
        print(i)
    obj = s3.get_object(Bucket='pronoo', Key=key)
    grid_sizes = pd.read_csv(obj['Body']).drop(['Unnamed: 0'], axis=1)
    all_players = pd.concat([all_players, grid_sizes], axis=0)
    
    
for i,key in enumerate(keysTournaments):
    if i%5000==0:
        print(i)
    obj = s3.get_object(Bucket='pronoo', Key=key)
    grid_sizes = pd.read_csv(obj['Body']).drop(['Unnamed: 0'], axis=1)
    all_tournaments = pd.concat([all_tournaments, grid_sizes], axis=0)

all_tournaments.to_csv('data/all_tournaments2020.csv', index=False)
all_players.to_csv('data/all_players2020.csv', index=False)    

all_players.to_csv('data/all_players2020.csv', index=False)    

pd.read_csv('data/all_tournaments.csv')

changer all tournement vers 2020 dans l'export, 
rajouter la condition 2020 dans la recherche de clés
